+++ interface/src/components/Gallery/GeneratedImageList.tsx
