+++ interface/src/components/Gallery/LoadingSpinner.tsx
